done and happiness has come)))

+++ src/main/java/mate/academy/internet/shop/model/Item.java
@@ -3,9 +3,9 @@ package mate.academy.internet.shop.model;
 public class Item {
     private Long id;
     private String name;
-    private Double price;
+    private Integer price;
 
-    public Item(String name, Double price) {
+    public Item(String name, Integer price) {
         this.name = name;
         this.price = price;
     }
@@ -26,11 +26,11 @@ public class Item {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
